requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
ldap3==2.9.1
PGPy==0.6.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyodbc==5.2.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-tds==1.16.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
wheel==0.45.1
SQLAlchemy==2.0.41